requirements.txt
africastalking==1.1.7
astroid==2.2.5
certifi==2019.6.16
chardet==3.0.4
contextlib2==0.5.5
Django==2.2.4
django-cors-headers==3.1.0
djangorestframework==3.10.1
djangorestframework-simplejwt==4.3.0
gunicorn==19.9.0
httpretty==0.9.6
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
pep8==1.7.1
psycopg2-binary==2.8.3
PyJWT==1.7.1
pylint==2.3.1
python-dotenv==0.10.3
pytz==2019.1
requests==2.22.0
schema==0.7.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.3
whitenoise==4.1.3
wrapt==1.11.2
